A whole bunch of changes, its now in a buildable state Everything is translated now, still have to verify all the types, offset etc

+++ src/LibBSP/Structs/Common/Lumps/NumList.ts
@@ -1,6 +1,6 @@
 import {ILump} from './ILump'
 import {BSP, LumpInfo, MapType} from '../../BSP/BSP'
-import {long} from '../../../../utils/number'
+import {int, long} from '../../../../utils/number'
 
 export enum DataType {
     Invalid = 0,
@@ -337,7 +337,7 @@ export class NumList implements ILump {
         return this.data
     }
 
-    public get(index: int): int {
+    public get(index: int): bigint {
         return this._backingArray[index]
     }
